Refactor: all datasets should be in vaex.datasets (vaex-core)
