+++ packages/vaex-core/vaex/__init__.py
@@ -613,37 +613,8 @@ def connect(url, **kwargs):
     from vaex.server import connect
     return connect(url, **kwargs)
 
-
-def example(dataset='astronomy'):
-    """Returns an example DataFrame which comes with vaex for testing/learning purposes.
-
-    Available datasets:
-    - 'astronomy': Sample data from Helmi & De Zeeuw 2000
-    - 'taxi': Sample data from the Yellow Taxi New York City dataset
-    - 'titanic': The Titanic passenger dataset
-    - 'iris': The Iris flower dataset
-
-    :param str dataset: Which dataset to load.
-    :rtype: DataFrame
-    """
-    if dataset == 'astronomy':
-        return vaex.datasets.helmi_de_zeeuw_10percent.fetch()
-    elif dataset == 'taxi':
-        return vaex.datasets.yellow_taxi_2012.fetch()
-    elif dataset == 'titanic':
-        return vaex.ml.datasets.load_titanic()
-    elif dataset == 'iris':
-        return vaex.ml.datasets.load_iris()
-    else:
-        raise ValueError(f'Can not find the "{dataset}"" dataset.')
-
-
-def zeldovich(dim=2, N=256, n=-2.5, t=None, scale=1, seed=None):
-    """Creates a zeldovich DataFrame.
-    """
-    import vaex.file
-    return vaex.file.other.Zeldovich(dim=dim, N=N, n=n, t=t, scale=scale)
-
+def example():
+    return vaex.datasets.astronomy()
 
 # create named logger, for all loglevels
 logger = logging.getLogger('vaex')
