+++ packages/vaex-ml/vaex/ml/catboost.py
@@ -26,7 +26,7 @@ class CatBoostModel(state.HasState):
 
     >>> import vaex
     >>> import vaex.ml.catboost
-    >>> df = vaex.ml.datasets.load_iris()
+    >>> df = vaex.datasets.iris()
     >>> features = ['sepal_width', 'petal_length', 'sepal_length', 'petal_width']
     >>> df_train, df_test = df.ml.train_test_split()
     >>> params = {
