+++ tests/ml/conftest.py
@@ -4,7 +4,7 @@ import vaex
 
 @pytest.fixture(scope='session')
 def df_iris_original():
-    return vaex.ml.datasets.load_iris()
+    return vaex.datasets.iris()
 
 
 @pytest.fixture(scope='function')
@@ -14,7 +14,7 @@ def df_iris(df_iris_original, df_factory):
 
 @pytest.fixture(scope='session')
 def df_iris_1e5_original():
-    return vaex.ml.datasets.load_iris_1e5()
+    return vaex.datasets.iris_1e5()
 
 
 @pytest.fixture(scope='function')
@@ -24,7 +24,7 @@ def df_iris_1e5(df_iris_1e5_original, df_factory):
 
 @pytest.fixture(scope='session')
 def df_titanic_original():
-    return vaex.ml.datasets.load_titanic()
+    return vaex.datasets.titanic()
 
 
 @pytest.fixture(scope='function')
