+++ tests/ml/xgboost_test.py
@@ -3,7 +3,6 @@ import numpy as np
 pytest.importorskip("xgboost")
 import xgboost as xgb
 import vaex.ml.xgboost
-import vaex.ml.datasets
 
 
 # the parameters of the model
